requirements.txt
pytest>=7.4,<9
selenium>=4.15,<5
allure-pytest>=2.13,<3
